requirements.txt
streamlit
textblob
pillow

